% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webhook.R
\name{whapi_flatten_webhook}
\alias{whapi_flatten_webhook}
\title{Flatten Whapi webhook payload into tidy rows for persistence (robust)}
\usage{
whapi_flatten_webhook(payload, verbose = TRUE)
}
\arguments{
\item{payload}{list parsed from JSON (e.g. plumber \code{req$body})}

\item{verbose}{logical, print logs with cli?}
}
\value{
tibble with normalized fields
}
\description{
Flatten Whapi webhook payload into tidy rows for persistence (robust)
}
