% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webhook.R
\name{whapi_to_num}
\alias{whapi_to_num}
\title{Safely coerce to numeric}
\usage{
whapi_to_num(x)
}
\arguments{
\item{x}{Any R object (usually scalar).}
}
\value{
A numeric scalar (\code{double}) or \code{NA_real_}.
}
\description{
Utility to coerce a value into numeric (\code{double}), returning \code{NA_real_} if
input is \code{NULL} or \code{NA}.
}
\examples{
whapi_to_num("123")
whapi_to_num(NULL)   # NA_real_

}
\keyword{internal}
